+++ packages/nakadachi-interactor-mixin-validator/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @rym-lib/nakadachi-interactor-mixin-validator
 
+## 1.2.9
+
+### Patch Changes
+
+- feat(query-module): add comprehensive test coverage and fix function-based rules implementation #54
+- Updated dependencies
+  - @rym-lib/exception@1.2.9
+  - @rym-lib/inversify-bundler@1.2.9
+  - @rym-lib/nakadachi@1.2.9
+
 ## 1.2.8
 
 ### Patch Changes